fix cbrt on win32
git-svn-id: http://svn.cern.ch/guest/lhcb/LHCb/trunk@47522 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 33bdbf361257b8311ce4a368677aea114a219dbd

+++ LHCb/Kernel/LHCbMath/LHCbMath/Power.h
@@ -71,7 +71,7 @@ namespace Gaudi
 #ifdef __GNUC__
       return ::cbrt ( __x           );
 #else
-      return ::pow  ( __x , 1.0/3.0 );
+      return ::pow  ( __x , (double)1.0/3.0 );
 #endif
     }  
 
@@ -92,7 +92,7 @@ namespace Gaudi
 #ifdef __GNUC__
       return ::cbrtf ( __x           );
 #else
-      return ::pow   ( __x , 1.0/3.0 );
+      return ::pow   ( __x , (double)1.0/3.0 );
 #endif
     }  
 
@@ -113,7 +113,7 @@ namespace Gaudi
 #ifdef __GNUC__
       return ::cbrtl ( __x           );
 #else
-      return ::pow   ( __x , 1.0/3.0 );
+      return ::pow   ( __x , (double)1.0/3.0 );
 #endif
     }
